updated the migration scripts
