+++ scripts/migration/deploy-factory.js
@@ -2,8 +2,8 @@ const fs = require("fs");
 const { ethers } = require("hardhat");
 const { printBalance } = require("../utils");
 
-const DEPLOYER_ADDRESS = "0x1a8a27A5AD3dE62719e65eEC79507218bF951E28";
-const DEPLOYER_NONCE = "0";
+const DEPLOYER_ADDRESS = "";
+const DEPLOYER_NONCE = "";
 
 const deployFactory = async () => {
   console.log("##################################################################");
@@ -22,19 +22,13 @@ const deployFactory = async () => {
     return;
   }
 
-  /**
-   * Factory Deployment
-   */
-
   const factory = await ethers.deployContract("BridgeFactory");
   await factory.waitForDeployment();
 
   const factoryAddress = await factory.getAddress();
   console.log(`Bridge factory address: ${factoryAddress}`);
 
-  /**
-   * Save the Factory address
-   */
+  // Save the Factory address
   const factoryAddresses = {
     factoryAddress: factoryAddress,
   };
