+++ tasks/templates/xerc20.json
